Move composable to separate file
